% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyticChromVAR.R
\name{computeDeviationsAnalytic}
\alias{computeDeviationsAnalytic}
\title{Analytic version of chromVAR's computeDeviations}
\usage{
computeDeviationsAnalytic(
  object,
  annotations,
  w = 0.1,
  bs = 50,
  bias = NULL,
  expectation = NULL
)
}
\arguments{
\item{object}{A SummarizedExperiment (or SingleCellExperiment) with an assay
'counts', or a count (sparse) matrix.}

\item{annotations}{Peak annotation (sparse) matrix, with motifs as columns,
or a SummarizedExperiment containing this in the first assay.}

\item{w}{Standard deviation of the Gaussian kernel. Values close to zero will
effectively mean that only peaks from the same bins are used as background,
which is suboptimal if the bin is sparsely populated. High values (e.g. >1)
will lead to homogeneous sampling, which will fail to correct for bias.
A value of 0.05 will be highly similar to the original chromVAR, and values
below 0.2 are recommended.}

\item{bs}{Number of bins per dimension (total bins = bs^2)}

\item{bias}{Per-peak bias (i.e. GC content). If omitted, will try to get it
from `rowData(object)$bias`.}

\item{expectation}{Optional vector of length equal to `nrow(object)`
giving the expected counts. If NULL, defaults to mean counts.}
}
\value{
A SummarizedExperiment containing the adjusted deviations and
  z-scores for each motif/sample.
}
\description{
Analytic version of chromVAR's computeDeviations
}
\author{
Pierre-Luc Germain
}
