% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertModel.R
\name{computeDeviationsWeighted}
\alias{computeDeviationsWeighted}
\title{computeDeviationsWeighted}
\usage{
computeDeviationsWeighted(counts, annotations, bgcounts = NULL, bg = NULL)
}
\arguments{
\item{counts}{A matrix of aggregated (weighted) counts, with motifs as rows
and samples as columns.}

\item{annotations}{A matrix of motif annotations, with peaks as rows and
motifs as columns.}

\item{bgcounts}{A summarizedExperiment object containing an assay with of background-weighted counts ('counts'), with peaks as rows
and samples as columns. Additionally needs to contain GC content (named `bias`) (and optional fragment length information, named `flbias`) in the rowData.}

\item{bg}{An optional `bcvBackground` object. Either `bg` or `bgcounts` must
be given.}
}
\value{
A list with two matrices, for deviations and z-scores.
}
\description{
computeDeviationsWeighted
}
\author{
Pierre-Luc Germain
}
