% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterChromVAR.R
\name{CVnorm}
\alias{CVnorm}
\title{CVnorm: chromVAR-inspired ATAC-seq normalization}
\usage{
CVnorm(
  object,
  bias = NULL,
  grouping = NULL,
  smoothGrouping = grouping,
  bs = 50,
  w = 0.05,
  Z = FALSE,
  enforceZeros = TRUE
)
}
\arguments{
\item{object}{A matrix of counts, or a SummarizedExperiment-like object with
an assay named 'counts'.}

\item{bias}{Per-peak bias (i.e. GC content). If omitted, will try to get it
from `rowData(object)$bias`.}

\item{grouping}{Optional grouping for the baseline expectation (prevents 
bias toward more abundant groups).}

\item{smoothGrouping}{Optional grouping to determine correction strength. 
If bias is consistent within these groups, correction is reduced.}

\item{bs}{Number of bins per dimension (total bins = `bs^2`).}

\item{w}{Standard deviation of the Gaussian kernel for bin smoothing.}

\item{Z}{Logical; whether to return standardized residuals (Z-scores) 
instead of the (default) corrected counts.}

\item{enforceZeros}{Logical; whether to enforce that zero counts should 
remain zeroes after correction (ignored if `Z=TRUE`).}
}
\value{
If `object` is a matrix, then a matrix of corrected counts of the 
  same dimensions. If `object` is a SummarizedExperiment-like object, then
  the object is returned with an extra "corrected" assay.
}
\description{
Corrects ATAC peak counts by removing the effects of technical biases 
(GC/accessibility) using the chromVAR background binning approach and an 
optional variance-based bias shrinkage (inspired from the qsmooth package) 
to preserve group biological signal.
}
\references{
- Schep A.N., Wu B., Buenrostro J.D., Greenleaf W.J. (2017) chromVAR: 
    inferring transcription-factor-associated accessibility from 
    single-cell epigenomic data, Nature Methods, doi: 10.1038/nmeth.4401
  - Hicks SC, Okrah K, Paulson JN, Quackenbush J, Irizarry RA, Corrado
    Bravo H (2018). “Smooth quantile normalization.” Biostatistics 19 (2),
    doi: 10.1093/biostatistics/kxx028
}
\author{
Pierre-Luc Germain
}
