% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.filterFrags}
\alias{.filterFrags}
\title{filterFrags}
\usage{
.filterFrags(atacFrag, min = 30, max = 2000)
}
\arguments{
\item{min}{and @param max the minimum and maximum limit of fragment length}

\item{atacFrag:}{a list of data tables containing the fragments information}
}
\value{
a list of data tables of filtered fragments
}
\description{
remove fragments that are too short or too long
}
\author{
Jiayi Wang
}
