% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertModel.R, R/utils.R
\name{.genomicRangesMapping}
\alias{.genomicRangesMapping}
\title{Mapping & aggregating modalities with genomic coordinates to reference
coordinates.}
\usage{
.genomicRangesMapping(
  refRanges,
  assayTable,
  byCols = c("tf_uniprot_id", "cell_type_id"),
  seqNamesCol = "chr",
  startCol = "start",
  endCol = "end",
  scoreCol = NULL,
  calledInCol = NULL,
  aggregationFun = NULL,
  minoverlap = 1,
  BPPARAM = SerialParam()
)

.genomicRangesMapping(
  refRanges,
  assayTable,
  byCols = c("tf_uniprot_id", "cell_type_id"),
  seqNamesCol = "chr",
  startCol = "start",
  endCol = "end",
  scoreCol = NULL,
  calledInCol = NULL,
  aggregationFun = NULL,
  minoverlap = 1,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{refRanges}{GRanges object with reference coordinates}

\item{assayTable}{modality table to be mapped to the reference coordinates.
Needs to containing genomic coordinates (see args: seqNamesCol, startCol, endCol), and byCols.}

\item{byCols}{will be the columns /depths of the resulting matrix with dimension
ref coords x byCols (or ref coord x byCols[1] x byCols[2]).
ByCols can be for instance cell type labels and/or transcription factor names.}

\item{seqNamesCol}{name of the column in motif, atac and chIP-seq data.tables containing
the sequence information.}

\item{startCol}{name of the column in motif, atac and chIP-seq data.tables containing
the start coordinate.}

\item{endCol}{name of the column in motif, atac and chIP-seq data.tables containing
the end coordinate.}

\item{scoreCol}{name of the score column (e.g. motif matching scores, atac fragment counts)}

\item{aggregationFun}{function (e.g. mean, median, sum) used to aggregate
if multiple rows of the assayTable overlap a reference coordinate.}

\item{BPPARAM}{BiocParallel argument either SerialParam() or MulticoreParam(workers=n)}
}
\description{
Internal convencience function for mapping different modality scores with
cell type and further labels such as tfs to reference coordinates. The resulting
table will have dimension ref coords x byCols (or ref coord x byCols[1] x byCols[2]).
ByCols can be for instance cell type labels and/or transcription factor names.

Internal convencience function for mapping different modality scores with
cell type and further labels such as tfs to reference coordinates. The resulting
table will have dimension ref coords x byCols (or ref coord x byCols[1] x byCols[2]).
ByCols can be for instance cell type labels and/or transcription factor names.
}
\author{
Emanuel Sonder
}
