% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertModel.R
\name{.getBackgroundPeaks}
\alias{.getBackgroundPeaks}
\title{Function to get the background peak sets and strongest motif match scores in
each peak}
\usage{
.getBackgroundPeaks(se, genome, niterations)
}
\arguments{
\item{se}{a SummarizedExperiment object that containing original fragment count
on peak level}

\item{genome}{BSgenome object}

\item{niterations}{number of background peaks to sample}
}
\value{
a list containing background peak indices and a matrix containing the
strongest motif match score for peaks per sample
}
\description{
Function to get the background peak sets and strongest motif match scores in
each peak
}
