% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getGCContent}
\alias{.getGCContent}
\title{getGCcontent}
\usage{
.getGCContent(gr, genome)
}
\arguments{
\item{gr:}{a GRanges object}

\item{genome:}{a BSgenome object, the corresponding genome}
}
\value{
a GRanges objects with an additional metadata column gc that contains
GC content
}
\description{
To calculate the GC content of each fragment or peak
}
\author{
Jiayi Wang
}
