% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertModel.R
\name{.getMotifActivityScore}
\alias{.getMotifActivityScore}
\title{get the motif activity score}
\usage{
.getMotifActivityScore(backgroundPeaks, peakMatchScore, bgScore)
}
\arguments{
\item{backgroundPeaks}{a data.table containing background peak indices}

\item{peakMatchScore}{a list of matrix containing match score on peak level
for each motif}

\item{maxScorePeaks}{a matrix containing strongest motif match score for
peaks per sample}
}
\value{
a list of matrix containing motif activity score on peak level per sample
}
\description{
get the motif activity score
}
