% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertModel.R
\name{.getPeakMatchScore}
\alias{.getPeakMatchScore}
\title{Function to get, from the counts-per-match and for a motif of choice,
a matrix of the counts from the best motif match per peak (0 if none)}
\usage{
.getPeakMatchScore(matchScore, peakRange, seqCol = "seqnames", ...)
}
\arguments{
\item{matchScore}{a data table containing motif ranges, motif_id, sample
and match score (output by `.getInsertProfile`)}

\item{peakRange}{a GRange object containing the peak ranges}
}
\value{
a list of counts-per-peaks for motif X
}
\description{
Function to get, from the counts-per-match and for a motif of choice,
a matrix of the counts from the best motif match per peak (0 if none)
}
