% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.resizeRanges}
\alias{.resizeRanges}
\title{.resizeRanges}
\usage{
.resizeRanges(
  peakRanges,
  width = 300,
  fix = c("center", "start", "end", "summit"),
  ...
)
}
\arguments{
\item{peakRanges:}{a GRanges object of peak ranges}

\item{width:}{the re-defined size of each peak}

\item{fix:}{the fixed point for resizing}
}
\value{
a GRange object with resized ranges
}
\description{
Resize the peaks
}
\author{
Jiayi Wang
}
