% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sanityCheck}
\alias{.sanityCheck}
\title{.sanityCheck}
\usage{
.sanityCheck(atacFrag, ranges, type = c("peaks", "motifs"))
}
\arguments{
\item{atacFrag:}{a list of data.tables that contain the ranges of fragments}

\item{peakRanges:}{a GRange object that contains the ranges of peaks}

\item{motifRanges:}{a GRange object that contains the ranges of motifs,
check metadata columns
check seqnames to factor in datatable}
}
\description{
Sanity check to ensure the input arguments have the correct classes
}
