% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dtToGr}
\alias{dtToGr}
\title{Convert a data.table to a GRanges object}
\usage{
dtToGr(
  dt,
  seqCol = "seqnames",
  startCol = "start",
  endCol = "end",
  strandCol = "strand",
  stranded = FALSE,
  addMetaCols = TRUE
)
}
\arguments{
\item{x}{A data.frame or data.table with genomic coordinates.}
}
\value{
A GRanges object.
}
\description{
Convert a data.table to a GRanges object
}
\examples{
library(data.table)
dt <- data.table(
    seqnames = c("chr1", "chr1"),
    start = c(100, 200),
    end = c(150, 250)
)
gr <- dtToGr(dt)
gr
}
\author{
Emanuel Sonder
}
