% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitDeviations}
\alias{fitDeviations}
\title{fitDeviations}
\usage{
fitDeviations(
  dev,
  design,
  type = c("norm", "deviations", "z"),
  contrasts = NULL,
  robust = FALSE,
  ...
)
}
\arguments{
\item{dev}{A SummarizedExperiment of motif deviations, as produced for
instance by \code{\link[betterChromVAR]{betterChromVAR}}.}

\item{design}{The design matrix of the experiment, as produced for instance 
\code{\link[limma]{model.matrix}}.}

\item{type}{The type of assay to use, either 'norm' (size-normalized 
z-scores), 'deviations' (background-adjusted motif deviations), or 'z' 
(z-scores).}

\item{contrasts}{Optional contrasts, passed to 
\code{\link[limma]{contrasts.fit}}.}

\item{robust}{Logical, passed to \code{\link[limma]{eBayes}}; whether the 
fit of the priors should be robust (not recommended).}

\item{...}{Other parameters passed to \code{\link[limma]{lmFit}}.}
}
\value{
A fitted `MArrayLM` object, which can for instance be used with
  limma's \code{\link[limma]{topTable}} function to extract differential
  analysis results.
}
\description{
A wrapper to fit a model to motif deviations, using limma to estimate 
posterior variability from a trend with the number of matches.
}
\examples{
dd <- betterChromVAR::getDummyData()
dev <- betterChromVAR::betterChromVAR(dd$counts, dd$motifMatches)
dev$group <- rep(LETTERS[1:2], each=5)
design <- stats::model.matrix(~group, data=SummarizedExperiment::colData(dev))
fit <- fitDeviations(dev, design)
}
