% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterChromVAR.R
\name{getBinProbMatrix}
\alias{getBinProbMatrix}
\title{getBinProbMatrix}
\usage{
getBinProbMatrix(x, bias = NULL, w = 0.1, bs = 50, pseudo = 0)
}
\arguments{
\item{bias}{Vector of GCBias (default from object)}

\item{w}{Standard deviation of the Gaussian kernel)}

\item{bs}{Number of bins per dimension (total bins = bs^2)}

\item{pseudo}{Optional pseudocount to be added. This should not be needed 
with standard workflows.}

\item{object}{A chromVAR counts object}
}
\value{
a list with the slots `peak2bin` (which bin each peak belongs to) 
  and `binBinProbs` (the probability of a peak from a given bin being 
  selected as background for another).
}
\description{
Get chromVAR-like bin-bin background selection probabilities
}
\references{
Schep, A.N., Wu, B., Buenrostro, J.D., Greenleaf W.J. chromVAR: 
  inferring transcription-factor-associated accessibility from 
  single-cell epigenomic data Nature Methods (2017)
}
