% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertModel.R
\name{getInsertionProfiles}
\alias{getInsertionProfiles}
\alias{getWeightedInsertions}
\title{Tn5 insertion counting}
\usage{
getInsertionProfiles(
  atacData,
  motifRanges,
  margin = 200,
  shift = FALSE,
  minFrag = 30,
  maxFrag = 3000,
  calcProfile = TRUE,
  profiles = NULL,
  symmetric = FALSE,
  stranded = FALSE,
  subSample = NULL,
  simplified = FALSE,
  BPPARAM = SerialParam()
)

getWeightedInsertions(
  atacData,
  peakRanges,
  motifRanges,
  genome,
  resize = FALSE,
  width = 300,
  minFrag = 30,
  maxFrag = 3000,
  shift = FALSE,
  ...
)
}
\arguments{
\item{atacData}{Named list of [GenomicRanges::GRanges-class], [data.table::data.table], data.frames or paths to .bed /. bam files
containing ATAC-seq fragment coordinates (i.e. chr/seqnames, start, end and optionally a strand column). List names will be used as sample names.
If a single object is provided and it contains a column named "sample", insertion counts will be computed for each sample separately.}

\item{motifRanges}{[GenomicRanges::GRanges-class] object containing coordinates of motif-matches.
Needs to contain a metadata column `motif_id` if insertion profiles should be computed for several motifs separately.}

\item{margin}{Margin around motif-matches to consider for computing Tn5 insertion events}

\item{shift}{If Tn5 insertion bias should be considered (only if strand column is provided).}

\item{minFrag}{Minimum fragment length to be considered for counting insertions. Default is 30bp.}

\item{maxFrag}{Maximum fragment length to be considered for counting insertions. Default is}

\item{calcProfile}{If insertion footprint profiles should be computed.}

\item{profiles}{Pre-computed insertion footprint profile as obtained by this function when running with `calcProfile=TRUE`.
Needs to contain a column with relative positions wrt to the center of the motif match (termed "rel_pos") and weight column (termed "w").}

\item{symmetric}{If transcription factor footprint profiles should be symmetric around the motif matches. Only used if `calcProfile=TRUE`.}

\item{stranded}{If insertion footprint profiles should be computed taking strandedness of fragments into account.}

\item{subSample}{If fragments should be sub-sampled for speed-up.
Default is no sub-sampling, if a number is provided the fragments of each file/object provided will be subsampled to that number.}

\item{simplified}{If return should be simplified and be provided as a [SummarizedExperiment::RangedSummarizedExperiment-class] object.}

\item{BPPARAM}{Parallel back-end to be used. Passed to [BiocParallel::bpmapply()].}

\item{peakRanges}{[GenomicRanges::GRanges-class] object containing coordinates of ATAC-seq peaks. Used for computing background weighted insertion counts.}

\item{resize}{If TRUE, peaks will be resized to the provided width around their center. Default is FALSE.}

\item{width}{If `resize=TRUE`, width to which peaks will be resized.}

\item{...}{Additional arguments passed to `getInsertionProfiles()`, e.g. for computing insertion profiles.}
}
\value{
If `simplified=TRUE` a [data.table::data.table] containing insertion counts within and in margins around motif matches and weighted insertion counts in case
an insertion profile is provided or if `calcProfile=TRUE`.
If `calcProfile=TRUE` also a footprint profile around the motif matches is returned, containing a weight ("w") column corresponding to relative insertion
frequency at the respective position relative to the motif center ("rel_pos").
If `simplified=FALSE` a [SummarizedExperiment::RangedSummarizedExperiment-class] object is returned with ranges corresponding to the `motifRanges` provided as an input.
Assays correspond to the different insertion counts computed, columns will correspond to the samples. If `calcProfile=TRUE` the profiles will be saved in the metadata.

A list containing the following elements: SummarizedExperiment object with weighted insertion counts for the motif matches (insertioncounts), a SummarizedExperiment object with weighted insertion counts for the peaks (bgcounts) and the insertion profile used for weighting (bgPeakProfile). If `calcProfile=TRUE` also a data.table with the computed insertion profiles per motif is returned (insertion_profiles).
}
\description{
Counts Tn5 insertions around provided motif-matches.
If requested also computes insertion footprint profiles and weighted insertion counts.
}
\author{
Emanuel Sonder
}
