% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifInteractions.R
\name{getInteractionsDeviations}
\alias{getInteractionsDeviations}
\title{getInteractionsDeviations}
\usage{
getInteractionsDeviations(se, annotation, bait, minCount = 20, ...)
}
\arguments{
\item{se}{An object inheriting `SummarizedExperiment`, containing a `counts`
assay.}

\item{annotation}{A (sparse) matrix or object inheriting 
`SummarizedExperiment` containing motif annotations for each row of `se`.}

\item{bait}{The name of the bait motif (corresponding to a column of 
`annotation`), for which to search for interacting motifs.}

\item{minCount}{The minimum overlap size for a motif pair to be considered.}

\item{...}{Passed to `betterChromVAR`}
}
\value{
A `SummarizedExperiment` of deviations for the intersection of 
  all motifs with the bait motif.
}
\description{
Get \code{\link[betterChromVAR]{betterChromVAR}} deviations for overlaps 
between motifs and a bait motif.
}
\examples{
attach(betterChromVAR::getDummyData())
dev <- getInteractionsDeviations(counts, motifMatches, bait="motif1",
                                 minCount=5)
dev
}
