% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getpmoi}
\alias{getpmoi}
\title{Get motif matches}
\usage{
getpmoi(
  genome,
  peaks,
  spec = c("Homo sapiens", "Mus musculus"),
  minHits = 50L,
  seqStyle = c("ensembl", "NCBI", "UCSC"),
  keepTop = NULL,
  srcFolder,
  motifs = NULL,
  thresh = NULL
)
}
\arguments{
\item{genome}{A BSgenome respective to the genome used for the alignment.}

\item{spec}{Species. Either "Hsapiens" or "Mmusculus".}

\item{seqStyle}{Either "ensembl" or "UCSC" depending on the format of the peak file.}

\item{srcFolder}{Folder where to find scripts and important objects}

\item{peakpath}{The path to the merged ATAC-seq peaks from control and treatment conditions.}
}
\description{
Get motif matches
}
\author{
Pierre-Luc
}
