% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedInsertions2.R
\name{weightedInsertions}
\alias{weightedInsertions}
\title{weightedInsertions}
\usage{
weightedInsertions(
  lf,
  peaks,
  motifMatches,
  extension = 200L,
  shift = NULL,
  ncores = 1,
  verbose = TRUE,
  weightMode = c("sub", "asIs"),
  smooth = 1/30,
  symm = FALSE,
  retFull = FALSE,
  res = NULL
)
}
\arguments{
\item{lf}{A named vector of paths to files, either (indexed) bam files or
tabix-indexed fragment files.}

\item{peaks}{A GRanges of peaks.}

\item{motifMatches}{A GRanges of motif matches, with a motif_id metadata 
column.}

\item{extension}{How much to extend the profile around the peak center.}

\item{shift}{Logical; whether to shift fragment ends. If not specified,
shifting is performed for bam inputs.}

\item{ncores}{The number of threads to use, or a BiocParallelParam object.}

\item{verbose}{Logical; whether to print progress messages.}

\item{weightMode}{Either 'sub' (subtract a fraction of the minimum value in
the profile before computing weights) or 'asIs' (use counts as is for 
computing weights).}

\item{smooth}{Smoothing span for lowess smoothing. If `smooth=TRUE`, Tukey's
running median smoothing (3RS3R) is used.}

\item{symm}{Logical; whether weights should be made symmetrical.}

\item{retFull}{Logical; whether to return the full matrices of counts around
all matches.}

\item{res}{Eventual list of pre-computed counts around matches, bypassing 
reading data. This is meant for debugging/development.}
}
\value{
A list with the following slots:
\describe{
\item{\code{peakCounts}:}{A SummarizedExperiment object of insertion counts per peak.}
\item{\code{wInsCounts}:}{A matrix of weighted insertion counts per motif (rows) and sample (column).}
\item{\code{profiles}:}{A matrix of summed count at each position (row) around matches for each motif (column).}
\item{\code{weights}:}{A matrix of weights at each position (row) around matches for each motif (column).}
\item{\code{peakAnnotation}:}{A sparse logical matrix of motif (column) annotations for each peak (row).}
}
}
\description{
weightedInsertions
}
