% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterChromVAR.R
\name{betterChromVAR}
\alias{betterChromVAR}
\title{betterChromVAR}
\usage{
betterChromVAR(
  object,
  annotations,
  grouping = NULL,
  bias = NULL,
  expectation = NULL,
  verbose = TRUE,
  bs = 50,
  sigma = 1,
  nthreads = NULL,
  w = 0.05,
  shrinkage = c("none", "average", "smooth")
)
}
\arguments{
\item{object}{A SummarizedExperiment (or SingleCellExperiment) with an assay
'counts', or a count (sparse) matrix.}

\item{annotations}{Peak annotation (sparse) matrix, with motifs as columns,
or a SummarizedExperiment containing this in the first assay. Values 
should be either logical or between 0 and 1.}

\item{grouping}{An optional factor or vector coercible to a factor indicating
the groupings of the columns of `object`. This is optionally used to 1) 
compute the base expectation such that rare cell types are given as much 
weight as abundant ones, and 2) apply shrinkage (if `shrinkage!="none"`) 
on a per-grouping fashion. In single-cell data, the grouping can for 
instance be the interaction of samples and cell types.}

\item{bias}{Per-peak bias (i.e. GC content). If omitted, will try to get it
from `rowData(object)$bias`.}

\item{expectation}{Optional vector of length equal to `nrow(object)` 
giving the expected counts. If NULL, defaults to mean counts (eventually
grouped, see `grouping`).}

\item{bs}{Number of bins per dimension (total bins = `bs^2`).}

\item{sigma}{Sigma parameter for the 2D smoothing. Ignored unless 
`shrinkage="smooth"`.}

\item{nthreads}{Either an integer scalar indicating the number of threads to
use, or a BiocParallelParam object. This is ignored unless `grouping` is 
given and shrinkage is applied.}

\item{w}{Standard deviation of the Gaussian kernel. Values close to zero will
effectively mean that only peaks from the same bins are used as background,
which is suboptimal if the bin is sparsely populated. High values (e.g. >1)
will lead to homogeneous sampling, which will fail to correct for bias.
A value of 0.05 will be highly similar to the original chromVAR, and values
below 0.2 are recommended.}
}
\value{
A SummarizedExperiment containing the adjusted deviations and 
  z-scores for each motif/sample.
}
\description{
A fast, analytic implementation of chromVAR's computeDeviations, 
additionally enabling balanced expectations and bias shrinkage.
}
\references{
Schep, A.N., Wu, B., Buenrostro, J.D., Greenleaf W.J. chromVAR: 
  inferring transcription-factor-associated accessibility from 
  single-cell epigenomic data Nature Methods (2017)
}
\author{
Pierre-Luc Germain
}
