% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.matchSeqlevels}
\alias{.matchSeqlevels}
\title{.matchSeqlevels}
\usage{
.matchSeqlevels(atacFrag, ranges)
}
\arguments{
\item{atacFrag:}{a list of data tables containing the fragments information}

\item{ranges:}{a genomic object of peaks/motifs}
}
\value{
A list with elements `atacFrag` (the input list, filtered to
seqnames common to both `atacFrag` and `ranges`) and `ranges` (the input
ranges, filtered the same way).
}
\description{
match the chromosomes between fragments and peaks/motifs
}
\author{
Jiayi Wang
}
