% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sanityCheck}
\alias{.sanityCheck}
\title{.sanityCheck}
\usage{
.sanityCheck(atacFrag, ranges, type = c("peaks", "motifs"))
}
\arguments{
\item{atacFrag:}{a list of data.tables that contain the ranges of fragments}

\item{ranges:}{a GRanges object that contains the ranges of peaks or motifs}

\item{type:}{either "peaks" or "motifs"; determines which sanity checks
are applied (e.g. for "motifs", checks that a `motif` metadata column
is present)}
}
\value{
NULL, invisibly; called for its side effect of raising an error
if the inputs are malformed.
}
\description{
Sanity check to ensure the input arguments have the correct classes
}
