% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dtToGr}
\alias{dtToGr}
\title{dtToGr}
\usage{
dtToGr(
  dt,
  seqCol = "seqnames",
  startCol = "start",
  endCol = "end",
  strandCol = "strand",
  stranded = FALSE,
  addMetaCols = TRUE
)
}
\description{
Convert a data table to GenomicRange object
}
\author{
Emanuel Sonder
}
