% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genomicRangesMapping}
\alias{genomicRangesMapping}
\title{Mapping & aggregation of modalities with genomic coordinates to a set of reference
coordinates.}
\usage{
genomicRangesMapping(
  refRanges,
  assayTable,
  byCols = c("tf_name", "cellular_context"),
  scoreCol = NULL,
  aggregationFun = NULL,
  minoverlap = 1,
  type = c("any", "start", "end", "within", "equal"),
  shift = FALSE,
  chunk = NULL,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{refRanges}{GRanges object with reference coordinates}

\item{assayTable}{List of [GenomicRanges::GRanges-class], [data.table::data.table], data.frames or paths to .bed /. bam files
containing assay data (e.g. fragments, motif scores, peaks) to be aggregated across the reference ranges provided.
Need to contain genomic coordinates (e.g. a chr/seqnames, start and end column).}

\item{byCols}{Variables across which the assays will be aggregated.
Will be the columns of the resulting [Matrix::Matrix-class]. If byCols is a vector with two elements,
the first one will constitute the list elements, the second the columns of the matrices being the list elements.
If they are factors (preferred) aggregation will happen across all levels, otherwise across all unique entries.}

\item{scoreCol}{name of the score column (e.g. motif matching scores, atac fragment counts) to be aggregated.
If it is NULL, the number of assay ranges overlapping the reference ranges will be counted.}

\item{aggregationFun}{function (e.g. mean, median, sum) used to aggregate.
If it is NULL, the number of assay ranges overlapping the reference ranges will be counted.}

\item{minoverlap}{Minimal overlap between refRanges and the assay ranges
Passed to [GenomicRanges::findOverlaps()]}

\item{type}{Type of overlap to be used, can be one of "any", "start", "end", "within", "equal".
Passed to [GenomicRanges::findOverlaps()]. Default is "any".}

\item{shift}{If Tn5 insertion bias should be considered (only if strand column is provided).}

\item{chunk}{If data should be processed in chunks (determined by chromosomes in `refRanges`). Recommended for large data.}

\item{BPPARAM}{Parallel back-end to be used. Passed to [BiocParallel::bplapply()].}
}
\value{
[Matrix::Matrix-class] or list of Matrices with rows corresponding to the reference ranges and columns (and list elements) to byCols.
}
\description{
Convencience function for mapping different modality scores with
cell type and further labels such as tfs to reference coordinates. The resulting
table will have dimension ref coords x byCols (or ref coord x byCols`[1]` x byCols`[2]`).
ByCols can be for instance cell type labels and/or transcription factor names.
}
\author{
Emanuel Sonder
}
