% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyticChromVAR.R
\name{getBinProbMatrix}
\alias{getBinProbMatrix}
\title{Get chromVAR bin-bin background selection probabilities}
\usage{
getBinProbMatrix(x, bias = NULL, w = 0.1, bs = 50)
}
\arguments{
\item{bias}{Vector of GCBias (default from object)}

\item{w}{Standard deviation of the Gaussian kernel)}

\item{bs}{Number of bins per dimension (total bins = bs^2)}

\item{object}{A chromVAR counts object}
}
\value{
a list with the slots `peak2bin` (which bin each peak belongs to) and
  `binBinProbs` (the probability of a peak from a given bin being selected
  as background for another).
}
\description{
Get chromVAR bin-bin background selection probabilities
}
